+++ core/src/hooks.cpp
@@ -1,10 +1,12 @@
 #include "core/hooks.hpp"
+#include "core/game.hpp"
 #include "core/string.hpp"
 
 #include "core/log.hpp"
 #include "core/state.hpp"
 #include "core/xinput.hpp"
 
+#include <libloaderapi.h>
 #include <windows.h>
 #include <MinHook.h>
 
@@ -15,12 +17,17 @@ using GetProcAddress_t = FARPROC(WINAPI*)(HMODULE hModule, LPCSTR lpProcName);
 GetProcAddress_t g_origGetProcAddress{};
 
 HMODULE WINAPI detourLoadLibraryExW(LPCWSTR lpLibFileName, HANDLE hFile, DWORD dwFlags) {
-    std::filesystem::path libfileName{lpLibFileName};
-    log("LoadLibraryExW: ", charStringFromChar8String(libfileName.u8string()));
-    if (libfileName.filename() == "gamedll_ph_x64_rwe.dll") {
+    std::filesystem::path libFileName{lpLibFileName};
+    log("LoadLibraryExW: ", charStringFromChar8String(libFileName.u8string()));
+    if (libFileName.filename() == "gamedll_ph_x64_rwe.dll") {
         log("Game DLL is being loaded.");
         auto module{g_origLoadLibraryExW(lpLibFileName, hFile, dwFlags)};
         AppState::get().notifyOnGameDllLoaded();
+        return module;
+    } else if (libFileName.filename() == "D3D12Core.dll") {
+        log("Redirecting loading of D3D12Core.");
+        const auto newPath{getMainBinDir() / libFileName.filename()};
+        auto module{g_origLoadLibraryExW(newPath.c_str(), hFile, dwFlags)};
         return module;
     }
     return g_origLoadLibraryExW(lpLibFileName, hFile, dwFlags);
@@ -30,7 +37,8 @@ FARPROC WINAPI detourGetProcAddress(HMODULE hModule, LPCSTR lpProcName) {
     auto modulePath{getModulePath(hModule)};
     if (hModule == AppState::get().getDllHandle()) {
         auto fixedModulePath{getXinput13Path()};
-        log("Shenanigan: GetProcAddress was used on us. Redirecting to module: ", charStringFromChar8String(fixedModulePath.u8string()));
+        // Temporarily disabled due to spam
+        //log("Shenanigan: GetProcAddress was used on us. Redirecting to module: ", charStringFromChar8String(fixedModulePath.u8string()));
         hModule = std::bit_cast<HMODULE>(getModuleHandle(fixedModulePath));
         return g_origGetProcAddress(hModule, lpProcName);
     }